feature add EmojiPickerInternalUtils and EmojiPickerExtendedUtils
