+++ lib/emoji_picker_flutter.dart
@@ -5,4 +5,4 @@ export 'package:emoji_picker_flutter/src/config.dart';
 export 'package:emoji_picker_flutter/src/emoji.dart';
 export 'package:emoji_picker_flutter/src/emoji_picker.dart';
 export 'package:emoji_picker_flutter/src/emoji_picker_builder.dart';
-export 'package:emoji_picker_flutter/src/emoji_picker_utils.dart';
+export 'package:emoji_picker_flutter/src/emoji_picker_extended_utils.dart';
